+++ PlaylistDJ.Frontend/src/pages/Home.svelte
