+++ PlaylistDJ.Frontend/src/pages/index.ts
@@ -1,9 +1,9 @@
 import About from './About.svelte'
-import HomeLoggedIn from './HomeLoggedIn.svelte'
+import Home from './Home.svelte'
 import NotFound from './NotFound.svelte'
 import Playlist from './Playlist.svelte'
 import PlaylistList from './PlaylistList.svelte'
 import Settings from './Settings.svelte'
 import Social from './Social.svelte'
 
-export { About, HomeLoggedIn, NotFound, Playlist, PlaylistList, Settings, Social }
+export { About, Home, NotFound, Playlist, PlaylistList, Settings, Social }
